VMS-14490: Input event is ignored on HIKVISION DS-2TD1217-6/V1. Review fix
VMS-14540: Fix review issues: naming, constants, literals, style
VMS-14540: UpdateInformationRestHandler refactoring: code simplifying, getting rid of redundant calls & queries
VMS-14540: std::function --> nx::utils::MoveOnlyFunc
VMS-14697: docker support: Enchance virtual environment (runtime) detection
VMS-14812: Add logging of audio/video device names when open usb cam
VMS-14813: Switch to new ffmpeg artifact, that had been built with antideadlock patch
VMS-14821. Review fix
VMS-14821: Offline status is constantly reoccurring for DS-2CD2132F-IWS camera when running RCT
VMS-14832: Fix review issues: indentation, quotes, warnings
VMS-14854: Fix input monitoring subscriptions in EventProcessor
VMS-14861: Review fix
VMS-14861: Review fix: `ignoreMedia2` renamed to `onvifIgnoreMedia2`
VMS-14864: Close http connection after every request, due to camera does not support multiple connections
VMS-14864: Fixed possible segfault
VMS-14867: Fix possible assert in HanwhaChunkLoader if channel chunks not found
VMS-14871: Use random::QtDevice instead of std::random_device to avoid ARM problems
VMS-14903: Server crashes on stop if ONVIF input port monitoring was enabled
VMS-14989: Fix root-tool.service requirement on cmake -DwithRootTool=OFF
VMS-14990: Keep mediaserver.conf on dpkg updates

<==== Mercurial data ====>
Branch: vms_4.0

+++ vms/server/nx_vms_server/src/plugins/resource/onvif/onvif_resource.cpp
@@ -810,13 +810,13 @@ CameraDiagnostics::Result QnPlOnvifResource::initOnvifCapabilitiesAndUrls(
         return CameraDiagnostics::CameraInvalidParams("ONVIF media URL is not filled by camera");
 
     const auto doIgnoreMedia2byPropertyKeyAsString = getProperty(
-        ResourcePropertyKey::Onvif::kIgnoreMedia2);
+        ResourcePropertyKey::kOnvifIgnoreMedia2);
     const bool doIgnoreMedia2byPropertyKey = doIgnoreMedia2byPropertyKeyAsString.toInt() > 0;
 
     const auto doIgnoreMedia2byDataKey = resourceData().value<bool>(
-        ResourceDataKey::kIgnoreMedia2, false);
+        ResourceDataKey::kOnvifIgnoreMedia2, false);
 
-    if (! (doIgnoreMedia2byPropertyKey || doIgnoreMedia2byDataKey))
+    if (!(doIgnoreMedia2byPropertyKey || doIgnoreMedia2byDataKey))
     {
         QString media2ServiceUrl;
         fetchOnvifMedia2Url(&media2ServiceUrl); //< We ignore the result,
@@ -3801,6 +3801,9 @@ void QnPlOnvifResource::stopInputPortStatesMonitoring()
         }
     }
 
+    if (serverModule()->isStopping())
+        return;
+
     if (QnSoapServer::instance() && QnSoapServer::instance()->getService())
         QnSoapServer::instance()->getService()->removeResourceRegistration(toSharedPointer(this));
 
@@ -4315,7 +4318,8 @@ bool QnPlOnvifResource::RenewSubscriptionAsOdmThreadSafe()
 
     _oasisWsnB2__Renew request;
 
-    auto interval = resourceData().value<QString>("renewIntervalForPullingAsOdm", "PT2M");
+    auto interval = resourceData().value<QString>(ResourceDataKey::kRenewIntervalForPullingAsOdm,
+        "PT2M");
     std::string odmRenewTerminationTime = interval.toStdString();
     request.TerminationTime = &odmRenewTerminationTime;
